+++ chessnotes/lib/main.dart
@@ -1,12 +1,11 @@
-
+import 'package:flutter/material.dart';
 import 'package:chessnotes/constants/routes.dart';
-import 'package:chessnotes/firebase_options.dart';
 import 'package:chessnotes/views/login_view.dart';
+import 'package:chessnotes/views/notes_view.dart';
 import 'package:chessnotes/views/register_view.dart';
 import 'package:chessnotes/views/verify_email_view.dart';
-import 'package:firebase_auth/firebase_auth.dart';
-import 'package:firebase_core/firebase_core.dart';
-import 'package:flutter/material.dart';
+import 'package:chessnotes/services/auth/auth_service.dart';
+
 
 void main() {
   WidgetsFlutterBinding.ensureInitialized();
@@ -33,15 +32,13 @@ class HomePage extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     return FutureBuilder(
-        future: Firebase.initializeApp(
-          options: DefaultFirebaseOptions.currentPlatform,
-        ),
+        future: AuthService.firebase().initialize(),
         builder: (context, snapshot) {
           switch (snapshot.connectionState) {
             case ConnectionState.done:
-              final user = FirebaseAuth.instance.currentUser;
+              final user = AuthService.firebase().currentUser;
               if (user != null) {
-                if (user.emailVerified) {
+                if (user.isEmailVerified) {
                   return const NotesView();
                 } else {
                   return const VerifyEmailView();
@@ -57,79 +54,7 @@ class HomePage extends StatelessWidget {
   }
 }
 
-enum MenuAction { logout }
 
-class NotesView extends StatefulWidget {
-  const NotesView({super.key});
-
-  @override
-  State<NotesView> createState() => _NotesViewState();
-}
-
-class _NotesViewState extends State<NotesView> {
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      appBar: AppBar(
-        title: const Text('Main UI'),
-        actions: [
-          PopupMenuButton<MenuAction>(
-            onSelected: (value) async {
-              switch (value) {
-                case MenuAction.logout:
-                  final shouldLogout = await showLogOutDialog(context);
-                  if (shouldLogout) {
-                    await FirebaseAuth.instance.signOut();
-                    Navigator.of(context).pushNamedAndRemoveUntil(
-                      loginRoute, 
-                      (_) => false,
-                    );
-                  }
-              }
-            },
-            itemBuilder: (context) {
-              return const [
-                PopupMenuItem<MenuAction>(
-                  value: MenuAction.logout,
-                  child: Text('Log out'),
-                ),
-              ];
-            },
-          )
-        ],
-      ),
-      body: const Text('Hello world'), 
-    );
-  }
-}
-
-
-
-Future<bool> showLogOutDialog(BuildContext context) {
-  return showDialog(
-    context: context, 
-    builder: (context) {
-      return AlertDialog(
-        title: const Text('Sign out'),
-        content: const Text('Are you sure you want to sign out?'),
-        actions: [
-          TextButton(
-            onPressed: () {
-              Navigator.of(context).pop(false);
-            }, 
-            child: const Text('Cancel'),
-          ),
-          TextButton(
-            onPressed: () {
-              Navigator.of(context).pop(true);
-            }, 
-            child: const Text('Log out'),
-          ),
-        ]
-      );
-    }
-  ).then((value) => value ?? false);
-}
