+++ chessnotes/lib/services/auth/auth_service.dart
@@ -1,9 +1,12 @@
 import 'package:chessnotes/services/auth/auth_user.dart';
 import 'package:chessnotes/services/auth/auth_provider.dart';
+import 'package:chessnotes/services/auth/firebase_auth_provider.dart';
 
 class AuthService implements AuthProvider {
   final AuthProvider provider;
   const AuthService(this.provider);
+
+  factory AuthService.firebase() => AuthService(FirebaseAuthProvider());
 
   @override
   Future<AuthUser> createUser({
@@ -33,4 +36,7 @@ class AuthService implements AuthProvider {
 
   @override
   Future<void> sendEmailVerification() => provider.sendEmailVerification();
+  
+  @override
+  Future<void> initialize() => provider.initialize();
 }
