+++ chessnotes/lib/services/auth/firebase_auth_provider.dart
@@ -1,3 +1,5 @@
+import 'package:chessnotes/firebase_options.dart';
+import 'package:firebase_core/firebase_core.dart';
 import 'package:chessnotes/services/auth/auth_user.dart';
 import 'package:chessnotes/services/auth/auth_provider.dart';
 import 'package:chessnotes/services/auth/auth_exceptions.dart';
@@ -5,6 +7,14 @@ import 'package:firebase_auth/firebase_auth.dart' show FirebaseAuth, FirebaseAut
 
 
 class FirebaseAuthProvider implements AuthProvider {
+    
+  @override
+  Future<void> initialize() async {
+    await Firebase.initializeApp(
+          options: DefaultFirebaseOptions.currentPlatform,
+        );
+  }
+  
   @override
   Future<AuthUser> createUser({
     required String email, 
@@ -92,5 +102,4 @@ class FirebaseAuthProvider implements AuthProvider {
       throw UserNotLoggedInAuthException();
     }
   }
-
 }
