+++ chessnotes/lib/views/register_view.dart
@@ -1,7 +1,8 @@
-import 'package:chessnotes/constants/routes.dart';
-import 'package:chessnotes/utilities/show_error_dialog.dart';
-import 'package:firebase_auth/firebase_auth.dart';
 import 'package:flutter/material.dart';
+import 'package:chessnotes/constants/routes.dart';
+import 'package:chessnotes/services/auth/auth_service.dart';
+import 'package:chessnotes/utilities/show_error_dialog.dart';
+import 'package:chessnotes/services/auth/auth_exceptions.dart';
 
 class RegisterView extends StatefulWidget {
   const RegisterView({super.key});
@@ -12,7 +13,7 @@ class RegisterView extends StatefulWidget {
 
 class _RegisterViewState extends State<RegisterView> {
   late final TextEditingController _email;
-  late final TextEditingController _password; 
+  late final TextEditingController _password;
 
   @override
   void initState() {
@@ -31,7 +32,8 @@ class _RegisterViewState extends State<RegisterView> {
   @override
   Widget build(BuildContext context) {
     return Scaffold(
-      appBar: AppBar(title: const Text('Register'),
+      appBar: AppBar(
+        title: const Text('Register'),
       ),
       body: Column(
         children: [
@@ -54,57 +56,47 @@ class _RegisterViewState extends State<RegisterView> {
             ),
           ),
           TextButton(
-            onPressed:() async {
-              
-      
+            onPressed: () async {
               final email = _email.text;
               final password = _password.text;
-              try{ 
-                await FirebaseAuth.instance.createUserWithEmailAndPassword(
-                  email: email, 
+              try {
+                await AuthService.firebase().createUser(
+                  email: email,
                   password: password,
                 );
-                final user = FirebaseAuth.instance.currentUser;
-                await user?.sendEmailVerification();
+                AuthService.firebase().sendEmailVerification();
                 Navigator.of(context).pushNamed(verifyEmailRoute);
-              } on FirebaseAuthException catch (e) {
-                if (e.code == 'weak-password') {
+              } on WeakPasswordException {
                   await showErrorDialog(
-                    context, 
+                    context,
                     'Weak password',
                   );
-                } else if (e.code == 'email-already-in-use') {
+              } on EmailAlreadyInUseException {
                   await showErrorDialog(
-                    context, 
+                    context,
                     'Email is already in use',
                   );
-                } else if (e.code == 'invalid-email') {
+              } on InvalidEmailException {
                   await showErrorDialog(
-                    context, 
+                    context,
                     'Email is invalid',
                   );
-                } else {
+              } on GenericAuthException {
                   await showErrorDialog(
-                    context, 
-                    'Error: ${e.code}',
+                    context,
+                    'Failed to register',
                   );
-                }
-              } catch (e) {
-                await showErrorDialog(
-                  context, 
-                  e.toString(),
-                );
-              }               
-            }, 
-              child: const Text('Register'),
+              }
+            },
+            child: const Text('Register'),
           ),
           TextButton(
             onPressed: () {
               Navigator.of(context).pushNamedAndRemoveUntil(
-              loginRoute, 
-              (route) => false,
-            );
-            }, 
+                loginRoute,
+                (route) => false,
+              );
+            },
             child: const Text('Already registered? Login here!'),
           ),
         ],
