+++ chessnotes/lib/views/verify_email_view.dart
@@ -1,6 +1,6 @@
-import 'package:chessnotes/constants/routes.dart';
-import 'package:firebase_auth/firebase_auth.dart';
 import 'package:flutter/material.dart';
+import 'package:chessnotes/constants/routes.dart';
+import 'package:chessnotes/services/auth/auth_service.dart';
 
 class VerifyEmailView extends StatefulWidget {
   const VerifyEmailView({super.key});
@@ -22,13 +22,12 @@ class _VerifyEmailViewState extends State<VerifyEmailView> {
           const Text("If you haven't received a verification email yet, press the button below."),
           TextButton(
             onPressed: () async {
-              final user = FirebaseAuth.instance.currentUser;
-              await user?.sendEmailVerification();
+              AuthService.firebase().sendEmailVerification();
             }, 
             child: const Text('Resend email verification')
           ),
           TextButton(onPressed: () async {
-            await FirebaseAuth.instance.signOut();
+            await AuthService.firebase().logOut();
             Navigator.of(context).pushNamedAndRemoveUntil(
               registerRoute, 
               (route) => false,
